MCOL-4671: MCOL-4622: fix the behavior of both PRs
first was playing different with RIGHT and LEFT functions(using the getUintVal and getIntVal accordingly)
https://github.com/mariadb-corporation/mariadb-columnstore-engine/pull/3234
second introduced round for ints from double, but added it to uint but not to int missing long doubles as well
https://github.com/mariadb-corporation/mariadb-columnstore-engine/pull/3480

+++ utils/funcexp/func_left.cpp
@@ -60,7 +60,7 @@ std::string Func_left::getStrVal(rowgroup::Row& row, FunctionParm& fp, bool& isN
   if (isNull || trimLength <= 0)
     return "";
 
-  size_t trimLengthPositive = static_cast<size_t>(trimLength);  // now we are sure it is positive
+  size_t trimLengthPositive = trimLength;  // now we are sure it is positive
   size_t charPos;
 
   if ((binLen <= trimLengthPositive) || (binLen <= (charPos = cs->charpos(pos, end, trimLengthPositive))))
